fix: Checkbox label no longer uses "for" syntax.
Refactored to wrap the <input> with a <label> element.  This fixes a bug where, if the id used for that component is not unique,
the wrong checkbox would be updated when clicking the label

Consider the following list:

li -> [checkbox id="foo"] [label for="foo"]
li -> [checbkox id="foo"] [label for="foo"]

Clicking the 2nd label would actually update the first checkbox's state.